moved isAdmin & isLoggedIn middleware to its own file in the server directory
