+++ server/middleware.js
@@ -1,0 +1,22 @@
+//middleware isAdmin
+const isAdmin = (req, res, next) => {
+  //check req.user.isAdmin is true
+  if (!req.user || !req.user.isAdmin) {
+    res.status(401).json({error: 'You are not authorized'})
+  }
+  next()
+}
+//middleware isLoggedIn
+const isLoggedIn = (req, res, next) => {
+  //check req.session.userId is there
+  if (!req.session.userId) {
+    res.status(401).json({error: 'You are not logged in'})
+    //redirect to log in page
+  }
+  next()
+}
+
+module.exports = {
+  isAdmin,
+  isLoggedIn
+}
